Debug Game Play and Improve Code
Move setting of next Turn to the end of checking Board's state.
Set board's state based on played card and not previous (did this
because previous will not yet be  the just played card, as
it will be set only when Turn is executed) (this was the error source).

Make BoardState's constructor parameters to be named and optional.
Get rid of clone factory and copyWith methods on BoardState.
Add assertions for Board's rule to BoardState's constructor.
Assign blank BoardStates where necessary, after gameplay.
Move BoardState and BoardExceptions to their files.

Remove './' in relative imports of files in the same folder.
Sort imports.

Get rid of size getter in CardCollection and use length directly,
as CardCollection extends Iterable and has a length property by default.